@@ -13,12 +13,31 @@
 # OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 class AdminController < ApplicationController
-  before_filter :login_required
-  before_filter :admin_only
-  #before_filter :signed_login_required, only: [:abort_signed_session]
-  #before_filter :token_login_required, only: [:abort_token_session]
+  before_action :login_required
+  before_action :admin_only
+  # before_action :signed_login_required, only: [:abort_signed_session]
+  # before_action :token_login_required, only: [:abort_token_session]
 
   def index
 
+  end
+
+  def show
+
+  end
+
+
+  def edit
+    password = params[:password]
+    password_confirmation = params[:password_confirmation]
+    if password && password == password_confirmation && password.length >= 8
+      current_user.password = password
+      current_user.save!
+      flash.now[:alert] = "The password has been changed successfully!"
+      redirect_to admin_url, notice: "Password changed!"
+    else
+      flash.now[:alert] = "The passwords don't match or password contains less than 8 characters"
+      render "show"
+    end
   end
 end
