@@ -14,7 +14,7 @@
 
 class FormsController < AdminController
   before_action :set_form, only: [:edit, :update, :destroy]
-  skip_before_filter :verify_authenticity_token
+  skip_before_action :verify_authenticity_token
   include FormsHelper
   respond_to :html, :js
 
