@@ -35,8 +35,9 @@ module CourseEval
     # config.i18n.default_locale = :de
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    # config.active_record.raise_in_transactional_callbacks = true
     config.encoding = 'utf-8'
+    config.load_defaults 5.0
 
   end
 end
