@@ -15,14 +15,6 @@
 # This file should contain all the record creation needed to seed the database with its default values.
 # The data can then be loaded with the rake db:seed (or created alongside the db with db:setup).
 #
-# Examples:
-#
-#   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
-#   Mayor.create(name: 'Emanuel', city: cities.first)
-#
-#admin = User.create([
-#                        {first_name: 'Admin',last_name: 'Admin', email: 'admin@myadmin.com', is_student: 'false', is_teacher: 'false', is_admin: 'true', is_management: 'false', password: 'admin'},
-#                        {first_name: 'Admin2',last_name: 'Admin2', email: 'evaluare@litere.ro', is_student: 'false', is_teacher: 'false', is_admin: 'true', is_management: 'false', password: 'admin'},
-#                        {first_name: 'Prof',last_name: 'Prof', email: 'prof@myprof.com', is_student: 'false', is_teacher: 'true', is_admin: 'false', is_management: 'false', password: 'prof'}
-#
-#                    ])
+
+
+admin = FactoryGirl.create(:user)
